Update mwo_stats.py
better exception handling; debug logging; command-line options

+++ MWO/mwo_stats.py
@@ -19,11 +19,17 @@ This will write out your stats along with some additional calculated stuff to mw
 You can then open that in your spreadsheet application of choice and see all the pretty data
 """
 import os
+import sys
+import argparse
 import pprint
 
 from getpass import getpass
 from selenium import webdriver
 from selenium.common.exceptions import NoSuchElementException
+from selenium.common.exceptions import WebDriverException
+from urllib3.exceptions import NewConnectionError
+from urllib3.exceptions import MaxRetryError
+from urllib3.exceptions import ProtocolError
 
 BASE_URL = "https://mwomercs.com/"
 PROF_URL = BASE_URL + "profile"
@@ -33,7 +39,7 @@ WEAP_DATA = STATS_URL + "?type=weapon"
 MAPS_DATA = STATS_URL + "?type=map"
 MODE_DATA = STATS_URL + "?type=mode"
 
-# special mech names
+# mech list
 MECH_NAME = {
     "SPARKY": "GRF-1E",
     "ARES": "GRF-AR",
@@ -145,30 +151,45 @@ def login_to_mwo(ffox):
     Log in
     Return the browser object
     """
-    print("Enter user email: ", end="")
-    email = input()
-
-    passwd = getpass("Enter user password: ")
-
-    ffox.get(PROF_URL)
-
-    login_link = ffox.find_element_by_link_text("LOGIN")
-    login_link.click()
-
-    email_input = ffox.find_element_by_id("email")
-    email_input.send_keys(email)
-
-    psswd_input = ffox.find_element_by_id("password")
-    psswd_input.send_keys(passwd)
-    psswd_input.submit()
-
     try:
-        alert = ffox.find_element_by_class_name("alert alert-error").text.strip()
-        if "Invalid email/password" in alert:
-            print("Invalid email/password entered, please try again.")
-            login_to_mwo(ffox)
-    except NoSuchElementException:
-        print("Logged in!")
+        print("Enter user email: ", end="")
+        email = input()
+
+        passwd = getpass("Enter user password: ")
+    except KeyboardInterrupt:
+        print("Aborted by user during username/password entry...")
+        sys.exit()
+
+    else:
+        try:
+            ffox.get(PROF_URL)
+
+            login_link = ffox.find_element_by_link_text("LOGIN")
+            login_link.click()
+
+            email_input = ffox.find_element_by_id("email")
+            email_input.send_keys(email)
+
+            psswd_input = ffox.find_element_by_id("password")
+            psswd_input.send_keys(passwd)
+            psswd_input.submit()
+        except (
+            ProtocolError,
+            MaxRetryError,
+            NewConnectionError,
+            WebDriverException,
+        ):
+            print("Aborted while attempting login. Bye!")
+            ffox.quit()
+            ffox.close()
+            sys.exit()
+        try:
+            alert = ffox.find_element_by_class_name("alert alert-error").text.strip()
+            if "Invalid email/password" in alert:
+                print("Invalid email/password entered, please try again.")
+                login_to_mwo(ffox)
+        except NoSuchElementException:
+            debug(5, ["Logged in!"])
 
 
 def get_mechbays(ffox):
@@ -176,10 +197,14 @@ def get_mechbays(ffox):
     Retrieve the profile page
     Iterate over mechBay classed tags and return the dictionary of owned my_mechs
     """
+    global args
+
     ffox.get(PROF_URL)
+    debug(1, ["Retrieving data from:", PROF_URL])
     mech_bays = ffox.find_elements_by_class_name("mechBay")
     mech_bay = {}
     for bay in mech_bays:
+        debug(2, [bay.text.strip()])
         try:
             mech_bay[bay.text.strip()] += 1
         except KeyError:
@@ -194,13 +219,17 @@ def get_stat_data(url, ffox):
     Find the table of stats and store each row as a dictionary entry under the first column
     Return the dictionary of stats
     """
+    global args
+
     ffox.get(url)
+    debug(1, ["Retrieving data from:", url])
     my_stats = {}
     stat_list = [x.text.strip() for x in ffox.find_elements_by_xpath("//thead/tr/th")]
     for obj in ffox.find_elements_by_xpath("//tbody/tr"):
         idx = 0
         obj_name = ""
         for stat in obj.find_elements_by_tag_name("td"):
+            debug(3, [stat.text.strip()])
             if idx == 0:
                 obj_name = stat.text.strip()
                 my_stats[obj_name] = {}
@@ -215,23 +244,40 @@ def get_owner_data(mechbays, stats):
     """
     Check stats entries against mechbay contents to see if you currently own the mech
     """
+    global args
+    debug(1, ["Parsing mechbay data for stats"])
     for mech in stats.keys():
         bay = mech
+        debug(4, [mech])
         if mech in MECH_NAME.keys():
             bay = MECH_NAME[mech]
 
         bay = bay.split(" ")[-1]  # Get the end of the mech name
+        debug(2, [bay])
         stats[mech]["Owned"] = bool(bay in mechbays.keys())
 
     return stats
+
+
+def debug(lvl, msg):
+    """
+    Print debug messages if the debug level is appropriate
+    """
+    if args.verbose >= lvl:
+        print("Debug" + str(lvl), ":", " ".join(msg))
 
 
 def main():
     """
     Main processing
     """
+    global args
+
     # Set up the log location for the geckodriver so we can delete it later
-    geckolog = "/tmp/geckodriver.log"
+    geckolog = args.log
+    if args.log is None:
+        print("Something went very wrong! Unable to log to 'None'. Aborting.")
+        sys.exit()
 
     # Set up the webdriver instance here so we can close it no matter what happens
     browser = webdriver.Firefox(service_log_path=geckolog)
@@ -239,6 +285,13 @@ def main():
     try:
         # Prompt the user for email/password and log in
         login_to_mwo(browser)
+    except KeyboardInterrupt:
+        print("Aborted by user during login. Bye!")
+        browser.quit()
+        browser.close()
+        sys.exit()
+
+    try:
         # Retrieve mechbays (owned mechs)
         my_mechs = get_mechbays(browser)
         # Retrieve all available mech stats
@@ -251,26 +304,67 @@ def main():
         my_maps_stats = get_stat_data(MAPS_DATA, browser)
         # Retrieve game mode stats
         my_mode_stats = get_stat_data(MODE_DATA, browser)
-    except Exception as we:
-        print(we)
+
+        if args.format == "dump":
+            pprint.pprint(my_mechs)
+            pprint.pprint(my_mech_stats)
+            pprint.pprint(my_weap_stats)
+            pprint.pprint(my_maps_stats)
+            pprint.pprint(my_mode_stats)
+
+    except KeyboardInterrupt:
+        print("Aborted post-login by user. Bye!")
         browser.quit()
         browser.close()
-
+        sys.exit()
+    except (
+        MaxRetryError,
+        NewConnectionError,
+        WebDriverException,
+    ) as e:
+        print(e)
+        browser.quit()
+        browser.close()
     finally:
         browser.close()
         browser.quit()
 
-    os.unlink(geckolog)
+    if args.retain:
+        os.unlink(geckolog)
 
-    pprint.pprint(my_mechs)
-    pprint.pprint(my_mech_stats)
-    pprint.pprint(my_weap_stats)
-    pprint.pprint(my_maps_stats)
-    pprint.pprint(my_mode_stats)
+
+def parse_args(argv):
+    """
+    parse command-line arguments
+    """
+    parser = argparse.ArgumentParser(description="Gather and Process player statistics from MWO profile")
+    parser.add_argument("--verbose", "-v", action="count", default=0)
+    parser.add_argument(
+        "--log",
+        action="store",
+        help="Log file for geckodriver",
+        default="/tmp/geckodriver.log",
+    )
+    parser.add_argument(
+        "--retain",
+        action="store_true",
+        help="Retain the webdriver log after completing",
+        default=False,
+    )
+    parser.add_argument(
+        "--format",
+        action="store",
+        choices=["csv", "xml", "dump", "keys", "xls"],
+        default="dump",
+        help="Output format",
+    )
+
+    return parser.parse_args(argv)
 
 
 if __name__ == "__main__":
     try:
+        args = parse_args(sys.argv[1:])
         main()
     except KeyboardInterrupt:
         print("Aborting...")
